// src/core/models/lawyer.ts
export interface Lawyer {
  id: number;
  full_name: string;
  phone: string;
  email: string;
  age: string;
  experience: string;
  yearly_income: string;
  has_children: string;
  license_states: string;
  expiration_date: string;
  license_number: string;
  duplicate_with: string;
}